travis: convert to container infrastructure
This uses the new container infrastructure in order to use build caches. The disadvantage is that we require that everything is available from Hackage, and this also can not test building the CUDA backend. However, this will hopefully reduce build times and reduce the chances of the build timing out or running out of memory.