pagerank: evaluate argument arrays before passing to successive `runN`
Make sure the inputs to each step of the computation (a runN) is evaluated first. On the PTX backend this was causing a 'blocked indefinitely on MVar' error, probably related to the way we are now taking exclusive access to a GPU for the duration of a computation. We should improve the ergonomics of that in the ptx backend, bit this fixes it here as well.

Fixes AccelerateHS/accelerate#415